delete RN Modal implementation of gif select
